services/auth/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
itsdangerous==2.1.2
click==8.1.7
Jinja2==3.1.3
PyJWT==2.8.0

pytest==8.3.3
pytest-cov==5.0.0
